pyproject.toml
[project]
name = "health-care-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "click==8.2.1",
    "colorama==0.4.6",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "markdown-it-py==3.0.0",
    "mcp[cli]==1.9.4",
    "mdurl==0.1.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.0",
    "pygments==2.19.2",
    "pymysql==1.1.1",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "rich==14.0.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.3.6",
    "starlette==0.47.1",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uv==0.7.13",
    "uvicorn==0.34.3",
]
